# Speed improvement: preprocessing without using pandas