Achieved a performance improvement by inputting features directly into LightGBM, bypassing pandas DataFrame. This is particularly effective when processing large datasets on a GPU.

- Example 1: Inference on 10,000 names with GPU

Before: 107.008s  
After: 38.383s

- Example 2: Inference on 10,000 names without GPU

Before: 162.693s  
After: 146.861s